+++ src/app/components/archive-settings.js
@@ -15,6 +15,7 @@ import { inject as service } from '@ember/service';
 import I18n from 'onedata-gui-common/mixins/components/i18n';
 import notImplementedIgnore from 'onedata-gui-common/utils/not-implemented-ignore';
 import notImplementedWarn from 'onedata-gui-common/utils/not-implemented-warn';
+import { promiseObject } from 'onedata-gui-common/utils/ember/promise-object';
 
 export default Component.extend(I18n, {
   // do not use tag, because the layout is built by `modal` property
@@ -23,6 +24,7 @@ export default Component.extend(I18n, {
 
   i18n: service(),
   globalNotify: service(),
+  archiveManager: service(),
 
   /**
    * @override
@@ -56,6 +58,13 @@ export default Component.extend(I18n, {
   modal: undefined,
 
   /**
+   * Injected options for archive creation.
+   * @virtual optional
+   * @type {CreateArchiveOptions}
+   */
+  options: undefined,
+
+  /**
    * True if submit Promise is pending.
    * @type {Boolean}
    */
@@ -79,6 +88,68 @@ export default Component.extend(I18n, {
    */
   canSubmit: reads('isValid'),
 
+  /**
+   * Initialized/set by `updateBaseArchiveProxy` method.
+   * @type {PromiseObject<Utils.BrowsableArchive>}
+   */
+  baseArchiveProxy: null,
+
+  init() {
+    this._super(...arguments);
+    if (this.get('options.baseArchive')) {
+      this.updateBaseArchiveProxy();
+    }
+  },
+
+  async getBaseArchive() {
+    const injectedBaseArchive = this.get('options.baseArchive');
+    if (injectedBaseArchive) {
+      return injectedBaseArchive;
+    } else {
+      try {
+        return await this.fetchLatestArchive();
+      } catch (error) {
+        // always resolve this promise, but pass error to form
+        console.debug(
+          `component:archive-settings#getBaseArchive: error getting baseArchive: ${error}`
+        );
+        return {
+          isCustomOnedataError: true,
+          type: 'cannot-fetch-latest-archive',
+          reason: error,
+        };
+      }
+    }
+  },
+
+  async updateBaseArchiveProxy() {
+    this.set('baseArchiveProxy', promiseObject(this.getBaseArchive()));
+  },
+
+  async fetchLatestArchive() {
+    const {
+      archiveManager,
+      dataset,
+    } = this.getProperties('archiveManager', 'dataset');
+    const archivesData = await archiveManager.fetchDatasetArchives({
+      datasetId: get(dataset, 'entityId'),
+      limit: 1,
+      offset: 0,
+    });
+    const archiveRecord = get(archivesData, 'childrenRecords.0');
+    if (archiveRecord) {
+      return archiveManager.getBrowsableArchive(get(archiveRecord, 'entityId'));
+    } else {
+      if (!archivesData.isLast) {
+        throw new Error(
+          'component:archive-settings#fetchLatestArchive: invalid archive listing data'
+        );
+      }
+      // there is no latest archive, because there are no archives
+      return null;
+    }
+  },
+
   async submitArchive() {
     const {
       formData,
@@ -88,7 +159,7 @@ export default Component.extend(I18n, {
     this.set('isSubmitting', true);
     try {
       if (canSubmit) {
-        const archiveCreateData = this.generateArchiveData(formData);
+        const archiveCreateData = await this.generateArchiveData(formData);
         return await onSubmit(archiveCreateData);
       }
     } finally {
@@ -96,7 +167,7 @@ export default Component.extend(I18n, {
     }
   },
 
-  generateArchiveData(formData) {
+  async generateArchiveData(formData) {
     if (formData) {
       const {
         config,
@@ -118,14 +189,16 @@ export default Component.extend(I18n, {
         'includeDip',
       );
       const isIncremental = Boolean(get(config, 'incremental'));
-      const baseArchiveId = get(config, 'baseArchiveId');
-      const incrementalConfig = {
-        enabled: isIncremental,
-      };
-      if (isIncremental && baseArchiveId) {
-        incrementalConfig.basedOn = baseArchiveId;
+      if (isIncremental) {
+        const baseArchive = this.get('baseArchiveProxy.content');
+        const baseArchiveId = baseArchive && get(baseArchive, 'entityId');
+        const incrementalConfig = {
+          enabled: isIncremental,
+          basedOn: baseArchiveId,
+        };
+        rawConfig.incremental = incrementalConfig;
       }
-      rawConfig.incremental = incrementalConfig;
+
       return {
         config: rawConfig,
         description,